laurel_sprout: Use speed tuning for performance critical applications
The default ART setting is quicken: run DEX code verification and optimize some DEX instructions to get better interpreter performance.

This commit switches performance critical applications to speed:
run DEX code verification and AOT-compile all methods.

PRODUCT_DEXPREOPT_SPEED_APPS (New in Android Oreo)
List of applications that have been identified as core to the products and which are desirable to compile with the speed compiler filter.  For example, persistent apps such as SystemUI get a chance to use profile-guided compilation only at the next reboot, so it may be better for the product to have these apps always AOT-compiled.